Cleanup encryption
