+++ src/Illuminate/Encryption/Encrypter.php
@@ -3,12 +3,20 @@
 namespace Illuminate\Encryption;
 
 use RuntimeException;
+use Illuminate\Support\Str;
 use Illuminate\Contracts\Encryption\DecryptException;
 use Illuminate\Contracts\Encryption\EncryptException;
 use Illuminate\Contracts\Encryption\Encrypter as EncrypterContract;
 
-class Encrypter extends BaseEncrypter implements EncrypterContract
+class Encrypter implements EncrypterContract
 {
+    /**
+     * The encryption key.
+     *
+     * @var string
+     */
+    protected $key;
+
     /**
      * The algorithm used for encryption.
      *
@@ -27,7 +35,7 @@ class Encrypter extends BaseEncrypter implements EncrypterContract
     {
         $key = (string) $key;
 
-        if (static::supported($key, $cipher)) {
+        if ($this->supported($key, $cipher)) {
             $this->key = $key;
             $this->cipher = $cipher;
         } else {
@@ -42,7 +50,7 @@ class Encrypter extends BaseEncrypter implements EncrypterContract
      * @param  string  $cipher
      * @return bool
      */
-    public static function supported($key, $cipher)
+    protected function supported($key, $cipher)
     {
         $length = mb_strlen($key, '8bit');
 
@@ -57,7 +65,7 @@ class Encrypter extends BaseEncrypter implements EncrypterContract
      */
     public function encrypt($value)
     {
-        $iv = openssl_random_pseudo_bytes($this->getIvSize());
+        $iv = openssl_random_pseudo_bytes(16);
 
         $value = openssl_encrypt(serialize($value), $this->cipher, $this->key, 0, $iv);
 
@@ -71,6 +79,18 @@ class Encrypter extends BaseEncrypter implements EncrypterContract
         $mac = $this->hash($iv = base64_encode($iv), $value);
 
         return base64_encode(json_encode(compact('iv', 'value', 'mac')));
+    }
+
+    /**
+     * Create a MAC for the given value.
+     *
+     * @param  string  $iv
+     * @param  string  $value
+     * @return string
+     */
+    protected function hash($iv, $value)
+    {
+        return hash_hmac('sha256', $iv.$value, $this->key);
     }
 
     /**
@@ -95,12 +115,56 @@ class Encrypter extends BaseEncrypter implements EncrypterContract
     }
 
     /**
-     * Get the IV size for the cipher.
+     * Get the JSON array from the given payload.
      *
-     * @return int
+     * @param  string  $payload
+     * @return array
+     *
+     * @throws \Illuminate\Contracts\Encryption\DecryptException
      */
-    protected function getIvSize()
+    protected function getJsonPayload($payload)
     {
-        return 16;
+        $payload = json_decode(base64_decode($payload), true);
+
+        // If the payload is not valid JSON or does not have the proper keys set we will
+        // assume it is invalid and bail out of the routine since we will not be able
+        // to decrypt the given value. We'll also check the MAC for this encryption.
+        if (!$payload || $this->invalidPayload($payload)) {
+            throw new DecryptException('The payload is invalid.');
+        }
+
+        if (!$this->validMac($payload)) {
+            throw new DecryptException('The MAC is invalid.');
+        }
+
+        return $payload;
+    }
+
+    /**
+     * Verify that the encryption payload is valid.
+     *
+     * @param  mixed  $data
+     * @return bool
+     */
+    protected function invalidPayload($data)
+    {
+        return !is_array($data) || !isset($data['iv']) || !isset($data['value']) || !isset($data['mac']);
+    }
+
+    /**
+     * Determine if the MAC for the given payload is valid.
+     *
+     * @param  array  $payload
+     * @return bool
+     *
+     * @throws \RuntimeException
+     */
+    protected function validMac(array $payload)
+    {
+        $bytes = Str::randomBytes(16);
+
+        $calcMac = hash_hmac('sha256', $this->hash($payload['iv'], $payload['value']), $bytes, true);
+
+        return Str::equals(hash_hmac('sha256', $payload['mac'], $bytes, true), $calcMac);
     }
 }
