+++ src/Illuminate/Encryption/EncryptionServiceProvider.php
@@ -2,7 +2,6 @@
 
 namespace Illuminate\Encryption;
 
-use RuntimeException;
 use Illuminate\Support\ServiceProvider;
 
 class EncryptionServiceProvider extends ServiceProvider
@@ -17,17 +16,7 @@ class EncryptionServiceProvider extends ServiceProvider
         $this->app->singleton('encrypter', function ($app) {
             $config = $app->make('config')->get('app');
 
-            $key = $config['key'];
-
-            $cipher = $config['cipher'];
-
-            if (Encrypter::supported($key, $cipher)) {
-                return new Encrypter($key, $cipher);
-            } elseif (McryptEncrypter::supported($key, $cipher)) {
-                return new McryptEncrypter($key, $cipher);
-            } else {
-                throw new RuntimeException('No supported encrypter found. The cipher and / or key length are invalid.');
-            }
+            return new Encrypter($config['key'], $config['cipher']);
         });
     }
 }
